Organize acceptance tests per public method
